+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All important changes to `fiisoft-jackdaw-stream` will be documented in this fil
 
 ## 1.1.0 - incompatible changes!
 
+- removed method Result::__toString
 - removed method ResultItem::create
 - added method StreamApi::collect
 - added method StreamApi::aggregate